Add -V to usage message.
Define new option -C to disable color even if available.  This and -t are not
implemented yet, and thus are not mentioned in the usage or command line.

Add a new -D, USE_COLOR to the Makefile, and turn it on by default.

ESR's color code will now be revived.